Moved commandlist alloc/free functions to CommandGroup
