+++ legion/engine/llri/detail/command_group.hpp
@@ -16,6 +16,7 @@ namespace LLRI_NAMESPACE
     enum struct result : uint8_t;
     class CommandList;
     enum struct queue_type : uint8_t;
+    struct command_list_alloc_desc;
 
     /**
      * @brief Describes how the CommandGroup should be created.
@@ -33,7 +34,7 @@ namespace LLRI_NAMESPACE
     };
 
     /**
-     * @brief CommandGroups are responsible for allocating the memory required to record CommandLists. They **may** be used to allocate one or multiple CommandLists but never more than command_group_desc::count.
+     * @brief CommandGroups are responsible for allocating the memory required to record CommandLists. They are used to allocate one or multiple CommandLists but never more than command_group_desc::count.
      *
      * @note CommandGroups are not thread-safe. CommandLists allocated through the same CommandGroup **can not** be recorded from separate threads simultaneously. For multi-threaded recording, create at least one separate CommandGroup per thread to prevent this from becoming an issue.
     */
@@ -54,6 +55,65 @@ namespace LLRI_NAMESPACE
         */
         result reset();
 
+        /**
+         * @brief Allocate a CommandList. The resulting CommandList will be of the same queue_type as the the CommandGroup, and is a non-owning pointer.
+         *
+         * Allocated CommandLists **may** be free'd through CommandGroup::free(), but if that isn't done, the CommandGroup will free all of its CommandLists upon destruction. CommandGroup maintains ownership over all of its allocated CommandLists.
+         *
+         * @param desc The allocation description describes the command list's usage.
+         * @param cmdList A pointer to the resulting command list variable;
+         *
+         * @return Success upon correct execution of the operation.
+         * @return ErrorInvalidUsage if cmdList was nullptr.
+         * @return ErrorInvalidUsage if desc.usage was not a valid command_list_usage enum value.
+         * @return ErrorExceededLimit if the CommandGroup doesn't have enough CommandLists to allocate another.
+         * @return Implementation defined result values: ErrorOutOfHostMemory, ErrorOutOfDeviceMemory.
+        */
+        result allocate(const command_list_alloc_desc& desc, CommandList** cmdList);
+
+        /**
+         * @brief Allocate one or more CommandLists. The resulting CommandLists will be of the same queue_type as the CommandGroup, and are non-owning pointers.
+         *
+         * Allocated CommandLists **may** be free'd through CommandGroup::free(), but if that isn't done, the CommandGroup will free all of its CommandLists upon destruction. CommandGroup maintains ownership over all of its allocated CommandLists.
+         *
+         * @param desc The allocation description describes the command list's usage. Each CommandList is allocated with this same structure.
+         * @param count The number of CommandLists to allocate.
+         * @param cmdLists A pointer to the resulting command list vector.
+         *
+         * @return Success upon correct execution of the operation.
+         * @return ErrorInvalidUsage if cmdLists was nullptr.
+         * @return ErrorInvalidUsage if desc.usage was not a valid command_list_usage enum value.
+         * @return ErrorInvalidUsage if count was 0.
+         * @return ErrorExceededLimit if the CommandGroup doesn't have enough CommandLists to allocate count CommandLists.
+         * @return Implementation defined result values: ErrorOutOfHostMemory, ErrorOutOfDeviceMemory.
+        */
+        result allocate(const command_list_alloc_desc& desc, uint8_t count, std::vector<CommandList*>* cmdLists);
+
+        /**
+         * @brief Free the CommandList, which allows the CommandGroup to release all of the device memory allocated for the CommandList.
+         *
+         * @param cmdList The CommandList to be free'd and destroyed.
+         *
+         * @return Success upon correct execution of the operation.
+         * @return ErrorInvalidUsage if cmdList is nullptr.
+         * @return ErrorInvalidUsage if the cmdList didn't belong to the CommandGroup.
+        */
+        result free(CommandList* cmdList);
+
+        /**
+         * @brief Free an array of CommandLists, which allows the CommandGroup to release all of the device memory allocated for the CommandLists.
+         *
+         * @param numCommandLists The number of CommandLists to be free'd, *should** match the size of the cmdLists array.
+         * @param cmdLists An array of CommandGroups, **must** not be nullptr and **must** at least contain numCommandLists amount of CommandLists.
+         *
+         * @return Success upon correct execution of the operation.
+         * @return ErrorInvalidUsage if numCommandLists is 0.
+         * @return ErrorInvalidUsage if cmdLists is nullptr.
+         * @return ErrorInvalidUsage if any of the CommandLists in cmdLists is nullptr.
+         * @return ErrorInvalidUsage if any of the CommandLists in cmdLists doesn't belong to the CommandGroup.
+        */
+        result free(uint8_t numCommandLists, CommandList** cmdLists);
+
     private:
         //Force private constructor/deconstructor so that only create/destroy can manage lifetime
         CommandGroup() = default;
@@ -72,5 +132,10 @@ namespace LLRI_NAMESPACE
         std::list<CommandList*> m_cmdLists;
 
         result impl_reset();
+
+        result impl_allocate(const command_list_alloc_desc& desc, CommandList** cmdList);
+        result impl_allocate(const command_list_alloc_desc& desc, uint8_t count, std::vector<CommandList*>* cmdLists);
+        result impl_free(CommandList* cmdList);
+        result impl_free(uint8_t numCommandLists, CommandList** cmdLists);
     };
 }
